# ops/openapi/runtime-router-openapi.json
{"openapi":"3.1.0","info":{"title":"RAG Runtime Router","version":"0.3.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/v1/models":{"get":{"summary":"List Models","operationId":"list_models_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Models V1 Models Get"}}}}}}},"/admin/cache/invalidate":{"post":{"summary":"Invalidate Cache","operationId":"invalidate_cache_admin_cache_invalidate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Invalidate Cache Admin Cache Invalidate Post"}}}}}}},"/admin/v1/vllm/status":{"get":{"summary":"Admin Vllm Status","operationId":"admin_vllm_status_admin_v1_vllm_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Admin Vllm Status Admin V1 Vllm Status Get"}}}}}}},"/admin/v1/vllm/load":{"post":{"summary":"Admin Vllm Load","operationId":"admin_vllm_load_admin_v1_vllm_load_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VLLMLoadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Admin Vllm Load Admin V1 Vllm Load Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/v1/vllm/unload":{"post":{"summary":"Admin Vllm Unload","operationId":"admin_vllm_unload_admin_v1_vllm_unload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Admin Vllm Unload Admin V1 Vllm Unload Post"}}}}}}},"/admin/v1/vllm/params":{"get":{"summary":"Admin Vllm Params Get","operationId":"admin_vllm_params_get_admin_v1_vllm_params_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Admin Vllm Params Get Admin V1 Vllm Params Get"}}}}}},"post":{"summary":"Admin Vllm Params Set","operationId":"admin_vllm_params_set_admin_v1_vllm_params_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VLLMParamsUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Admin Vllm Params Set Admin V1 Vllm Params Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/chat/completions":{"post":{"summary":"Chat Completions","operationId":"chat_completions_v1_chat_completions_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"VLLMLoadRequest":{"properties":{"model":{"type":"string","title":"Model"},"host":{"type":"string","title":"Host","default":"127.0.0.1"},"port":{"type":"integer","title":"Port","default":8000},"tensor_parallel_size":{"type":"integer","minimum":1.0,"title":"Tensor Parallel Size","default":1},"gpu_memory_utilization":{"type":"number","maximum":0.99,"exclusiveMinimum":0.0,"title":"Gpu Memory Utilization","default":0.9},"max_model_len":{"type":"integer","minimum":512.0,"title":"Max Model Len","default":8192},"served_model_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Served Model Name"},"trust_remote_code":{"type":"boolean","title":"Trust Remote Code","default":true},"cuda_visible_devices":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cuda Visible Devices"},"extra_args":{"type":"string","title":"Extra Args","default":""}},"type":"object","required":["model"],"title":"VLLMLoadRequest"},"VLLMParamsUpdate":{"properties":{"temperature":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temperature"},"top_p":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Top P"},"top_k":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top K"},"max_tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Tokens"},"presence_penalty":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Presence Penalty"},"frequency_penalty":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Frequency Penalty"},"repetition_penalty":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Repetition Penalty"},"min_p":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min P"},"reset":{"type":"boolean","title":"Reset","default":false}},"type":"object","title":"VLLMParamsUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}